+++ Res/TaskExt.cs
@@ -13,7 +13,7 @@ namespace ComradeVanti.CSharpTools
             return f(v);
         }
 
-        internal static async Task<TMapped> Bind<TValue, TMapped>(this Task<TValue> task, Func<TValue, Task<TMapped>> f)
+        private static async Task<TMapped> Bind<TValue, TMapped>(this Task<TValue> task, Func<TValue, Task<TMapped>> f)
         {
             var v = await task;
             return await f(v);
@@ -28,7 +28,7 @@ namespace ComradeVanti.CSharpTools
         /// <typeparam name="TValue">The type of the tasks success-value</typeparam>
         /// <typeparam name="TError">The type of the error</typeparam>
         /// <returns>The new task</returns>
-        internal static async Task<Res<TValue, TError>> ToAsyncRes<TValue, TError>(
+        public static async Task<Res<TValue, TError>> ToAsyncRes<TValue, TError>(
             this Task<TValue> task,
             Func<Exception, TError> makeError)
         {
@@ -53,7 +53,7 @@ namespace ComradeVanti.CSharpTools
         /// <typeparam name="TError">The type of the results error</typeparam>
         /// <typeparam name="TMapped">The type of the mapped results success</typeparam>
         /// <returns>A task with the mapped result</returns>
-        internal static Task<Res<TMapped, TError>> MapAsyncRes<TValue, TError, TMapped>(
+        public static Task<Res<TMapped, TError>> MapAsyncRes<TValue, TError, TMapped>(
             this Task<Res<TValue, TError>> task,
             Func<TValue, TMapped> f) =>
             task.Map(res => res.Map(f));
@@ -67,7 +67,7 @@ namespace ComradeVanti.CSharpTools
         /// <typeparam name="TError">The type of the results error</typeparam>
         /// <typeparam name="TMapped">The type of the mapped results error</typeparam>
         /// <returns>A task with the mapped result</returns>
-        internal static Task<Res<TValue, TMapped>> MapAsyncResError<TValue, TError, TMapped>(
+        public static Task<Res<TValue, TMapped>> MapAsyncResError<TValue, TError, TMapped>(
             this Task<Res<TValue, TError>> task,
             Func<TError, TMapped> f) =>
             task.Map(res => res.MapError(f));
@@ -81,7 +81,7 @@ namespace ComradeVanti.CSharpTools
         /// <typeparam name="TError">The type of the results error</typeparam>
         /// <typeparam name="TMapped">The type of the mapped results success</typeparam>
         /// <returns>A task with the mapped result</returns>
-        internal static Task<Res<TMapped, TError>> BindAsyncRes<TValue, TError, TMapped>(
+        public static Task<Res<TMapped, TError>> BindAsyncRes<TValue, TError, TMapped>(
             this Task<Res<TValue, TError>> task,
             Func<TValue, Res<TMapped, TError>> f) =>
             task.Map(res => res.Bind(f));
@@ -95,7 +95,7 @@ namespace ComradeVanti.CSharpTools
         /// <typeparam name="TError">The type of the results error</typeparam>
         /// <typeparam name="TMapped">The type of the mapped results success</typeparam>
         /// <returns>A task with the mapped result</returns>
-        internal static Task<Res<TMapped, TError>> BindAsyncResAsync<TValue, TError, TMapped>(
+        public static Task<Res<TMapped, TError>> BindAsyncResAsync<TValue, TError, TMapped>(
             this Task<Res<TValue, TError>> task,
             Func<TValue, Task<Res<TMapped, TError>>> f) =>
             task.Bind(res => res.BindAsync(f));